fix(paging): handle empty count when no elements match the request
