+++ lib/methods/index.ts
@@ -1,1 +1,2 @@
-export { GetResult, GetPagingResult } from './get-result/get-result';
+export { GetPagingResult } from './get-paging-result/get-paging-result';
+export { GetResult } from './get-result/get-result';
